return correct propagated thread ID to receiver when propagator and receiver are local to one another
In the case of a propagated message, current->get_partner()'s result may not
refer to the same thread as from_tcb->get_{local,global}_id(). This changeset
narrows the case where local TIDs are returned to avoid this confusion.

This changeset also makes explicit (through the use of an ASSERT()) the
assumption that only global IDs are passed through propagation.